+++ examples/configs/src/view-configs/spatial-beta/spatialdata-visium_io.js
@@ -4,6 +4,7 @@ import {
   CoordinationLevel as CL,
   hconcat,
   vconcat,
+  getInitialCoordinationScopePrefix,
 } from '@vitessce/config';
 
 // Reference: https://github.com/giovp/spatialdata-sandbox/blob/3da0af016d3ddd85f1d63a9c03a67b240b012bd0/visium_io/download.py#L15
@@ -21,9 +22,6 @@ function generateVisiumIoConfig() {
     url: baseUrl,
     options: {
       path: 'images/Visium_Mouse_Olfactory_Bulb_full_image',
-    },
-    coordinationValues: {
-      fileUid: 'histology-image',
     },
   }).addFile({
     fileType: 'obsFeatureMatrix.spatialdata.zarr',
@@ -67,38 +65,25 @@ function generateVisiumIoConfig() {
   const obsSets = config.addView(dataset, 'obsSets');
   const featureList = config.addView(dataset, 'featureList');
 
-  const [featureSelectionScope] = config.addCoordination('featureSelection');
-  featureSelectionScope.setValue(['Atp1b1']);
+  // const [featureSelectionScope] = config.addCoordination('featureSelection');
+  // featureSelectionScope.setValue(['Atp1b1']);
 
   const [obsColorEncodingScope] = config.addCoordination('obsColorEncoding');
   obsColorEncodingScope.setValue('geneSelection');
 
   config.linkViewsByObject([spatialView, lcView], {
-    spatialTargetZ: 0,
-    spatialTargetT: 0,
-    obsType: 'spot', // TODO: remove this after auto-initialization is supported per-layer/per-layer-channel.
-    // For now, cheating by allowing the spotLayer to fall back to the auto-initialized values for the view.
-    imageLayer: CL({
-      fileUid: 'histology-image',
-      spatialLayerOpacity: 1,
-      spatialLayerVisible: true,
-      photometricInterpretation: 'RGB',
-    }),
     spotLayer: CL({
-      obsType: 'spot',
-      spatialLayerVisible: true,
-      spatialLayerOpacity: 0.5,
-      spatialSpotRadius: 50.0,
-      featureValueColormapRange: [0, 0.5],
+      // featureValueColormapRange: [0, 0.5],
       obsColorEncoding: obsColorEncodingScope,
-      featureSelection: featureSelectionScope,
+      // featureSelection: featureSelectionScope,
     }),
-  });
+  }, { scopePrefix: getInitialCoordinationScopePrefix('A', 'obsSpots') });
+  // or
 
-  config.linkViews([featureList, heatmap, obsSets], ['obsType'], ['spot']);
+  config.linkViews([featureList, heatmap, obsSets, spatialView, lcView], ['obsType'], ['spot']);
 
-  featureList.useCoordination(featureSelectionScope);
-  heatmap.useCoordination(featureSelectionScope);
+  // featureList.useCoordination(featureSelectionScope);
+  // heatmap.useCoordination(featureSelectionScope);
 
   featureList.useCoordination(obsColorEncodingScope);
   heatmap.useCoordination(obsColorEncodingScope);