+++ packages/config/src/VitessceConfig.js
@@ -1,6 +1,6 @@
 /* eslint-disable react-hooks/rules-of-hooks */
 import { CoordinationType } from '@vitessce/constants-internal';
-import { fromEntries, getNextScope } from '@vitessce/utils';
+import { fromEntries, getNextScope, createPrefixedGetNextScopeNumeric } from '@vitessce/utils';
 
 /**
  * Class representing a file within a Vitessce config dataset.
@@ -430,11 +430,12 @@ export class VitessceConfigCoordinationScope {
    * Construct a new coordination scope instance.
    * @param {string} cType The coordination type for this coordination scope.
    * @param {string} cScope The name of the coordination scope.
+   * @param {[any]} cValue Optional. The coordination value of the coordination scope.
    */
-  constructor(cType, cScope) {
+  constructor(cType, cScope, cValue = null) {
     this.cType = cType;
     this.cScope = cScope;
-    this.cValue = null;
+    this.cValue = cValue;
   }
 
   /**
@@ -555,6 +556,7 @@ export class VitessceConfig {
       layout: [],
       initStrategy: 'auto',
     };
+    this.getNextScope = getNextScope;
   }
 
   /**
@@ -569,7 +571,7 @@ export class VitessceConfig {
   addDataset(name = undefined, description = undefined, options = undefined) {
     const { uid } = options || {};
     const prevDatasetUids = this.config.datasets.map(d => d.dataset.uid);
-    const nextUid = (uid || getNextScope(prevDatasetUids));
+    const nextUid = (uid || this.getNextScope(prevDatasetUids));
     const newDataset = new VitessceConfigDataset(nextUid, name, description);
     this.config.datasets.push(newDataset);
     const [newScope] = this.addCoordination(CoordinationType.DATASET);
@@ -641,7 +643,7 @@ export class VitessceConfig {
           ? Object.keys(this.config.coordinationSpace[cType])
           : []
       );
-      const scope = new VitessceConfigCoordinationScope(cType, getNextScope(prevScopes));
+      const scope = new VitessceConfigCoordinationScope(cType, this.getNextScope(prevScopes));
       if (!this.config.coordinationSpace[scope.cType]) {
         this.config.coordinationSpace[scope.cType] = {};
       }
@@ -668,8 +670,8 @@ export class VitessceConfig {
         : []
     );
     const metaContainer = new VitessceConfigMetaCoordinationScope(
-      getNextScope(prevMetaScopes),
-      getNextScope(prevMetaByScopes),
+      this.getNextScope(prevMetaScopes),
+      this.getNextScope(prevMetaByScopes),
     );
     if (!this.config.coordinationSpace[CoordinationType.META_COORDINATION_SCOPES]) {
       this.config.coordinationSpace[CoordinationType.META_COORDINATION_SCOPES] = {};
@@ -777,6 +779,9 @@ export class VitessceConfig {
     */
     const processLevel = (level) => {
       const result = {};
+      if (level === null) {
+        return result;
+      }
       Object.entries(level).forEach(([cType, nextLevelOrInitialValue]) => {
         // Check if value of object is instanceof CoordinationLevel
         // (otherwise assume it is the coordination value).
@@ -859,10 +864,19 @@ export class VitessceConfig {
    * instance, or a `CoordinationLevel` instance.
    * The CL function takes an array of objects as its argument, and returns a CoordinationLevel
    * instance, to support nesting.
-   * @param {boolean} meta Should meta-coordination be used? Optional. By default, true.
+   * @param {object|null} options
+   * @param {bool} options.meta Should meta-coordination be used? Optional.
+   * By default, true.
+   * @param {string|null} options.scopePrefix A prefix to add to all
+   * coordination scope names. Optional.
    * @returns {VitessceConfig} This, to allow chaining.
    */
-  linkViewsByObject(views, input, meta = true) {
+  linkViewsByObject(views, input, options = null) {
+    const { meta = true, scopePrefix = null } = options || {};
+
+    if (scopePrefix) {
+      this.getNextScope = createPrefixedGetNextScopeNumeric(scopePrefix);
+    }
     const scopes = this.addCoordinationByObject(input);
     if (meta) {
       const metaScope = this.addMetaCoordination();
@@ -876,7 +890,28 @@ export class VitessceConfig {
         view.useCoordinationByObject(scopes);
       });
     }
+    if (scopePrefix) {
+      this.getNextScope = getNextScope;
+    }
     return this;
+  }
+
+  /**
+   * Set the value for a coordination scope.
+   * If a coordination object for the coordination type does not yet exist
+   * in the coordination space, it will be created.
+   * @param {string} cType The coordination type.
+   * @param {string} cScope The coordination scope.
+   * @param {any} cValue The initial value for the coordination scope.
+   * @returns {VitessceConfigCoordinationScope} A coordination scope instance.
+   */
+  setCoordinationValue(cType, cScope, cValue) {
+    const scope = new VitessceConfigCoordinationScope(cType, cScope, cValue);
+    if (!this.config.coordinationSpace[scope.cType]) {
+      this.config.coordinationSpace[scope.cType] = {};
+    }
+    this.config.coordinationSpace[scope.cType][scope.cScope] = scope;
+    return scope;
   }
 
   /**
@@ -972,4 +1007,24 @@ export class VitessceConfig {
     });
     return vc;
   }
+}
+
+// For usage during auto-initialization.
+export function getCoordinationSpaceAndScopes(partialCoordinationValues, scopePrefix) {
+  const vc = new VitessceConfig({ schemaVersion: '1.0.16', name: '__dummy__' });
+  vc.getNextScope = createPrefixedGetNextScopeNumeric(scopePrefix);
+  const dataset = vc.addDataset('__dummy__');
+  const v1 = vc.addView(dataset, '__dummy__');
+  vc.linkViewsByObject([v1], partialCoordinationValues, { meta: true });
+  const vcJson = vc.toJSON();
+  // TODO: remove the "dataset" coordination type from these objects.
+  const { coordinationSpace } = vcJson;
+  const { coordinationScopes } = vcJson.layout[0];
+  const { coordinationScopesBy } = vcJson.layout[0];
+
+  return {
+    coordinationSpace,
+    coordinationScopes,
+    coordinationScopesBy,
+  };
 }