+++ packages/schemas/src/file-def-options.ts
@@ -87,9 +87,19 @@ export const imageOmeTiffSchema = z.object({
     .optional(),
 });
 
+export const obsSegmentationsOmeTiffSchema = imageOmeTiffSchema.extend({
+  obsTypesFromChannelNames: z.boolean()
+    .optional(),
+});
+
 // OME-Zarr (NGFF)
 export const imageOmeZarrSchema = z.object({
   coordinateTransformations: omeCoordinateTransformations
+    .optional(),
+});
+
+export const obsSegmentationsOmeZarrSchema = imageOmeZarrSchema.extend({
+  obsTypesFromChannelNames: z.boolean()
     .optional(),
 });
 
@@ -101,6 +111,7 @@ export const imageSpatialdataSchema = imageOmeZarrSchema.extend({
 export const obsSegmentationsSpatialdataSchema = z.object({
   // TODO: should this also extend the imageOmeZarrSchema?
   // TODO: should this be renamed labelsSpatialdataSchema?
+  // TODO: support obsTypesFromChannelNames?
   path: z.string(),
 });
 export const obsLocationsSpatialdataSchema = z.object({