+++ packages/schemas/src/index.ts
@@ -25,6 +25,8 @@ export {
   imageOmeZarrSchema,
   imageOmeTiffSchema,
   imageSpatialdataSchema,
+  obsSegmentationsOmeTiffSchema,
+  obsSegmentationsOmeZarrSchema,
   obsSegmentationsSpatialdataSchema,
   obsFeatureMatrixSpatialdataSchema,
   obsSpotsSpatialdataSchema,