+++ packages/view-types/feature-list/src/FeatureListSubscriber.js
@@ -43,6 +43,8 @@ export function FeatureListSubscriber(props) {
     showTable = false,
     sort = 'alphabetical',
     sortKey = null,
+    closeButtonVisible,
+    downloadButtonVisible,
   } = props;
 
   const loaders = useLoaders();
@@ -108,6 +110,8 @@ export function FeatureListSubscriber(props) {
       // even though this no longer uses the TitleInfo component's
       // scroll css (SelectableTable is virtual scroll).
       isScroll
+      closeButtonVisible={closeButtonVisible}
+      downloadButtonVisible={downloadButtonVisible}
       removeGridComponent={removeGridComponent}
       isReady={isReady}
       urls={urls}