+++ packages/view-types/statistical-plots/src/CellSetExpressionPlotSubscriber.js
@@ -108,6 +108,8 @@ function useExpressionByCellSet(
 export function CellSetExpressionPlotSubscriber(props) {
   const {
     coordinationScopes,
+    closeButtonVisible,
+    downloadButtonVisible,
     removeGridComponent,
     theme,
   } = props;
@@ -188,6 +190,8 @@ export function CellSetExpressionPlotSubscriber(props) {
   return (
     <TitleInfo
       title={`Expression by ${capitalize(obsType)} Set${(firstGeneSelected ? ` (${firstGeneSelected})` : '')}`}
+      closeButtonVisible={closeButtonVisible}
+      downloadButtonVisible={downloadButtonVisible}
       removeGridComponent={removeGridComponent}
       urls={urls}
       theme={theme}