+++ CHANGELOG.md
@@ -1,4 +1,22 @@
 
+## 3.3.2
+
+### Patch Changes
+
+- Add the prop closeButtonVisible to all views, passed down to TitleInfo, to allow hiding the close buttons on each view by setting this prop in the config. (`@vitessce/vit-s`) ([#1740](https://github.com/vitessce/vitessce/pull/1740))
+
+- Memoize button icons in TitleInfo to prevent unnecessary re-renders" (`@vitessce/vit-s`) ([#1736](https://github.com/vitessce/vitessce/pull/1736))
+
+- Added a prop to hide download buttons and dropdowns. (`@vitessce/vit-s`) ([#1743](https://github.com/vitessce/vitessce/pull/1743))
+
+
+## 3.3.1
+
+### Patch Changes
+
+- Update Deck.gl to 8.8.27 and Luma.gl to 8.5.21 (`@vitessce/layer-controller-beta`, `@vitessce/spatial-beta`, `@vitessce/gl`, `demo`, `docs`, `html`) ([#1724](https://github.com/vitessce/vitessce/pull/1724))
+
+
 ## 3.3.0
 
 ### Minor Changes